refactor: seperated list-item into its own components
