+++ src/components/ListItem.tsx
@@ -1,0 +1,51 @@
+import styled from '@emotion/styled'
+import { TiMinus } from 'react-icons/ti'
+
+interface Props {
+  item: string
+  index: number
+}
+
+const Container = styled.div`
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  width: 381px;
+  height: 62px;
+  border-bottom: 1px solid #5b5b5b;
+  margin-top: 22px;
+  padding-left: 18px;
+  padding-right: 18px;
+  box-sizing: border-box;
+  flex-shrink: 0;
+`
+
+const Text = styled.p`
+  margin: 0;
+  padding: 0;
+  font-family: Inter;
+  font-style: normal;
+  font-size: 24px;
+  font-weight: 500;
+`
+
+const Icon = styled(TiMinus)`
+  margin: 0;
+  padding: 0;
+  color: #7f7f7f;
+  width: 40px;
+  height: 40px;
+  border: 2px solid #808080;
+  border-radius: 50%;
+`
+
+const ListItem = ({ item, index }: Props) => {
+  return (
+    <Container key={index}>
+      <Text>{item}</Text>
+      <Icon></Icon>
+    </Container>
+  )
+}
+
+export default ListItem
